Variety of cleanups
