+++ core/src/threadpool.cpp
@@ -77,59 +77,3 @@ int ThreadPool::nthreads()
 {
   return _nthreads;
 }
-
-template<class F, class... Args>
-auto ThreadPool::submit(F&& f, Args&&... args) 
-  -> std::future<typename std::result_of<F(Args...)>::type>
-{
-  using return_type = typename std::result_of<F(Args...)>::type;
-
-  auto task = std::make_shared< std::packaged_task<return_type()> >(
-    std::bind(std::forward<F>(f), std::forward<Args>(args)...)
-  );
-        
-  std::future<return_type> res = task->get_future();
-  {
-    std::unique_lock<std::mutex> lock(queue_mutex);  
-    
-    // don't allow enqueueing after stopping the pool
-    if(stop)
-      throw std::runtime_error("enqueue on stopped ThreadPool");  
-    tasks.emplace([task](){ (*task)(); });
-  }
-  condition.notify_one();
-  return res;
-}
-
-// add new work item to the pool
-template<class F, class... Args>
-auto ThreadPool::enqueue(F&& f, Args&&... args) 
-    -> typename std::result_of<F(Args...)>::type
-{
-  using return_type = typename std::result_of<F(Args...)>::type;
-
-  auto task = std::make_shared< std::packaged_task<return_type()> >(
-    std::bind(std::forward<F>(f), std::forward<Args>(args)...)
-  );
-        
-  std::future<return_type> res = task->get_future();
-  {
-    std::unique_lock<std::mutex> lock(queue_mutex);  
-    
-    // don't allow enqueueing after stopping the pool
-    if(stop)
-      throw std::runtime_error("enqueue on stopped ThreadPool");  
-    tasks.emplace([task](){ (*task)(); });
-  }
-  condition.notify_one();
-  return res.get();
-}
-
-// add new work item to the pool
-template<class F, class... Args>
-auto ThreadPool::critical(F&& f, Args&&... args) 
-    -> typename std::result_of<F(Args...)>::type
-{
-  std::unique_lock<std::mutex> lock(runtime_mutex);
-  return f(args...);
-}
