+++ src/hydra.cpp
@@ -1,4 +1,5 @@
 #include "hydra.h"
+#include "distributetask.h"
 
 void Hydra::run()
 {
